requirements.txt
torch>=2.0.0
transformers>=4.30.0
diffusers>=0.18.0
accelerate>=0.20.0
pillow>=9.5.0
numpy>=1.24.0
tqdm>=4.65.0
datasets>=2.12.0
timm>=0.9.0 